requirements.txt
py-clob-client-v2
python-dotenv>=1.0.1
keyboard>=0.13.5
httpx>=0.27.0
colorama>=0.4.6
websocket-client>=1.6.0
web3>=6.0.0
